--- HelloAdam.txt
@@ -1,2 +1,0 @@
-Hello Adam!
-This is going to create a merge conflict
